Add StackTemplate class to project
To describe an AWS CloudFormation Stack which is meant to be created
we decided to introduce a StackTemplate. A StackTemplate contains
the name of the stack to be created, the CloudFormation template or
a link to it, parameters and tags for the stack.

As the Stack class provided by the AWS sdk does not fulfill all our
needs to describe a stack which is meant to be deployed but not yet
deployed, we defined the StackTemplate as something different than
the stack itself.

The StackTemplate class comes along with tests, test resources and
updated dependencies in the build.gradle.kts. To avoid checking in
build outputs also the .gitignore file was updated.