Add MfaCredentialsProvider class to project
To assume a role in a Multi-Account setup while using MFA we creat a
custom AWSCredentialsProvider.

It comes along with a Integration test which only works if you have an
AWS profile configured on your machine.

While we keep adding a lot of classes in the beginning of this project
we decided to only update the Patch Version. With this commit it is
updated to 0.0.2-SNAPSHOT.