Make StackTemplate immutable.
So far, the state of a StackTemplate after applying one of the "withers" was not defined by the unit tests. In fact it was altered, which might be rather unexpected. With this commit we define it to be unaltered and change the implementation accordingly.

The version is updated to 0.0.3-SNAPSHOT.