Add CertificateClient for issuing ACM certificates.
We add the class CertificateClient which can request an ACM certificate for a given domain in a given hosted zone and automatically validate it via Route 53 domain validation. If the corresponding certificate already exists, it will return the ARN of the existing certificate.

The class comes with unit tests and this commit increments the verison to 0.0.4-SNAPSHOT.